requirements.txt
black==25.1.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
isort==6.0.0
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.1
RapidFuzz==3.12.1
six==1.17.0
thefuzz==0.22.1
tzdata==2025.1
